refactor: コンポーネントの命名規則を変更
- ついでにロゴタイプのフォントスタイルを変更
