--- components/header/logo.tsx
@@ -1,11 +1,0 @@
-export default function Logo() {
-  return (
-    <a class="flex items-center" href="/">
-      <i class="ti ti-at text-2xl md:text-4xl" />
-      <span class="ml-2 md:ml-3 text-sm md:text-base">
-        <p>im@sparql</p>
-        <p>Data Dashboard</p>
-      </span>
-    </a>
-  );
-}
